test(e2e): harmonize test fixtures
